Adding deserialization

+++ RTS/RTS/Component.cpp
@@ -2,7 +2,7 @@
 
 
 void ComponentContainer::registerComponent(Component* component) {
-	unsigned long id = component->componentId;
+	Uint8 id = component->componentId;
 	if (mComponents.find(id) != mComponents.end()) {
 		throw std::exception("Component type already registered.");
 	}
@@ -10,7 +10,7 @@ void ComponentContainer::registerComponent(Component* component) {
 	mComponents.emplace(id, component);
 }
 
-void ComponentContainer::deregisterComponent(unsigned long id) {
+void ComponentContainer::deregisterComponent(Uint8 id) {
 	if (mComponents.find(id) == mComponents.end()) {
 		throw std::exception("No component registered.");
 	}
@@ -18,6 +18,6 @@ void ComponentContainer::deregisterComponent(unsigned long id) {
 	mComponents.erase(id);
 }
 
-Component* ComponentContainer::getComponentByType(unsigned long type) {
+Component* ComponentContainer::getComponentByType(Uint8 type) {
 	return mComponents.at(type);
 }
